split feature and post-processing
